updated configs, added prod, added server response handlers

+++ src/server/method-request-handler.ts
@@ -2,69 +2,97 @@ import {IncomingMessage, ServerResponse} from 'http';
 import {UserDto} from '../models/models';
 import {db} from '../db/db-service';
 import {parseDataRequest} from './parse-data-request';
+import {getParamId, isValidUserRequiredData, uuidIsValid} from '../helpers/helpers';
+import {
+    invalidUserRequiredDataResponse,
+    invalidUserResponse,
+    invalidUuidResponse,
+    successUserCreateResponse,
+    successUserDeletedResponse,
+    successUserResponse,
+    successUsersResponse,
+    successUserUpdateResponse
+} from './response-handlers';
+
 
 export class MethodRequestHandler {
 
     static async getRequestHandler(req: IncomingMessage, res: ServerResponse) {
-        const userId = (req.url as string).split('/')[3];
+        const userId = getParamId(req);
         if (userId) {
+            if (!uuidIsValid(userId)) {
+                await invalidUuidResponse(res);
+                return;
+            }
             const user = await db.getUser(userId);
-            res.writeHead(200, { 'Content-Type': 'application/json' });
-            res.end(JSON.stringify(user));
+            if (!user) {
+                await invalidUserResponse(res);
+                return;
+            }
+            await successUserResponse(res, user);
             return;
         }
         const data = await db.getAllUsers();
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.write(JSON.stringify(data));
-        res.end();
+        await successUsersResponse(res, data);
     }
 
     static async postRequestHandler(req: IncomingMessage, res: ServerResponse) {
         try {
             const data = await parseDataRequest(req);
-            const newUser: UserDto = JSON.parse(data);
-            //
-            // if (!isUserDataValid(newUser)) {
-            //     await handleInvalidUserDataRequest(res);
-            //
-            //     return;
-            // }
 
+            if (!isValidUserRequiredData(data)) {
+                await invalidUserRequiredDataResponse(res);
+                return;
+            }
+            const newUser: UserDto = JSON.parse(data);
             const user = await db.createUser(newUser);
 
-            res.writeHead(201, {'Content-Type': 'application/json'});
-            res.end(JSON.stringify(user));
+            if (!user) {
+                await invalidUserResponse(res);
+                return;
+            }
+
+            await successUserCreateResponse(res, user);
         } catch (e) {
-            // if (e instanceof SyntaxError) {
-            //     await handleInvalidUserDataRequest(res);
-            // } else {
-            //     await handleInternalError(res);
-            // }
         }
     }
 
 
     static async putRequestHandler(req: IncomingMessage, res: ServerResponse) {
-        const userId = (req.url as string).split('/')[3];
+        const userId = getParamId(req);
         if (userId) {
+            if (!uuidIsValid(userId)) {
+                await invalidUuidResponse(res);
+                return;
+            }
             const data = await parseDataRequest(req);
             let updateUserFields: Partial<UserDto> = JSON.parse(data);
 
-
             const updatedUser = await db.updateUser(userId, updateUserFields);
-            res.writeHead(200, {'Content-Type': 'application/json'});
-            res.end(JSON.stringify(updatedUser));
+            if (!updatedUser) {
+                await invalidUserResponse(res);
+                return;
+            }
+            await successUserUpdateResponse(res, updatedUser);
         }
 
     }
 
     static async deleteRequestHandler(req: IncomingMessage, res: ServerResponse) {
-        const userId = (req.url as string).split('/')[3];
+        const userId = getParamId(req);
+
         if (userId) {
+            if (!uuidIsValid(userId)) {
+                await invalidUuidResponse(res);
+                return;
+            }
+            const user = await db.getUser(userId);
+            if (!user) {
+                await invalidUserResponse(res);
+                return;
+            }
             await db.deleteUser(userId);
-            res.writeHead(204, { 'Content-Type': 'application/json' });
-            res.end();
+            await successUserDeletedResponse(res);
         }
     }
